Move timeline from components/ to core/
