+++ web/src/beta/core/Crust/Widgets/Widget/Navigator/index.tsx
@@ -2,7 +2,7 @@ import type { ComponentProps as WidgetProps } from "..";
 import { Visible } from "../useVisible";
 
 import useHooks from "./hooks";
-import NavigatorPresenter from "./NavigatorPresenter";
+import NavigatorUI from "./UI";
 
 export type Props = WidgetProps<Property>;
 
@@ -42,7 +42,7 @@ const Navigator = ({
   });
 
   return visible ? (
-    <NavigatorPresenter theme={theme} degree={degree} editing={!!editing} {...events} />
+    <NavigatorUI theme={theme} degree={degree} editing={!!editing} {...events} />
   ) : null;
 };
